feat: update to version 1.3.4-beta.5 with battery entity filtering improvements
- Added a new utility function `getBatteryEntities` to streamline battery entity filtering in `SummaryCard` and `BatteriesViewStrategy`.
- Refactored battery entity logic to utilize binary and percentage sensors only.
