+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "simon42-dashboard-strategy",
-  "version": "1.3.4-beta.4",
+  "version": "1.3.4-beta.5",
   "private": true,
   "type": "module",
   "scripts": {
